Pull the process data submission function into a helper
It's not very coupled to the submit page, and will be useful for save
and return.
